refactor: 공통으로 사용되는 컴포넌트를 common 폴더로 이동
adminEdit 페이지와 myPage에서 공통으로 사용되는 컴포넌트가 존재
myPage에 생성해뒀던 컴포넌트 중 adminEdit 페이지에서 공통으로 사용되는 것을 common 폴더로 이동
(추후, signup 페이지와 공통으로 사용되는 부분도 체크해야합니다.)
