+++ src/common/components/atom/accountManagement/subHeader/MainLogoStyles.ts
