+++ src/common/components/molecules/accountManagement/subHeader/SubHeader.tsx
@@ -1,0 +1,17 @@
+import MainLogo from "~/common/components/atom/accountManagement/subHeader/MainLogo";
+import SubTitle from "~/common/components/atom/accountManagement/subHeader/SubTitle";
+import { SubHeaderBox } from "./SubHeaderStyles";
+
+interface Props {
+  subTitle: string;
+}
+const SubHeader = ({ subTitle }: Props) => {
+  return (
+    <SubHeaderBox>
+      <MainLogo />
+      <SubTitle text={subTitle} />
+    </SubHeaderBox>
+  );
+};
+
+export default SubHeader;
